# docs/v2/customer/customer_swagger.yaml
definitions:
  customer.EntitiesResponseTemplate-customer_bankAccountResp:
    properties:
      count:
        type: integer
      next:
        type: string
      previous:
        type: string
      results:
        items:
          $ref: '#/definitions/customer.bankAccountResp'
        type: array
    type: object
...
  customer.EntitiesResponseTemplate-customer_transactionResp:
    properties:
      count:
        type: integer
      next:
        type: string
      previous:
        type: string
      results:
        items:
          $ref: '#/definitions/customer.transactionResp'
        type: array
    type: object
  customer.bankAccountResp:
    properties:
      account_number:
        type: string
      cash_in:
        type: number
      cash_out:
        type: number
      create_time:
        type: string
      customer_id:
        type: string
      id:
        type: string
      is_for_payment:
        type: boolean
      update_time:
        type: string
    type: object
  customer.bankAccountUpdateReq:
    properties:
      is_for_payment:
        type: boolean
    required:
    - is_for_payment
    type: object
...
  customer.errorResponse:
    properties:
      code:
        type: string
      detail:
        type: string
      message:
        type: string
    type: object
...
  customer.meResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      username:
        type: string
    type: object
...
  customer.transactionCreateReq:
    properties:
      amount:
        type: number
      description:
        type: string
      is_fee_paid_by_me:
        type: boolean
      receiver_id:
        type: string
    required:
    - amount
    - description
    - is_fee_paid_by_me
    - receiver_id
    type: object
  customer.transactionCreateResp:
    properties:
      amount:
        type: number
      create_time:
        type: string
      description:
        type: string
      id:
        type: string
      receiver_bank_account_number:
        type: string
      receiver_bank_name:
        type: string
      receiver_id:
        type: string
      receiver_name:
        type: string
      sender_bank_account_number:
        type: string
      sender_bank_name:
        type: string
      sender_id:
        type: string
      sender_name:
        type: string
      source_transaction_id:
        type: string
      status:
        $ref: '#/definitions/transaction.Status'
      token:
        type: string
      transaction_type:
        $ref: '#/definitions/transaction.TransactionType'
      update_time:
        type: string
    type: object
  customer.transactionResp:
    properties:
      amount:
        type: number
      create_time:
        type: string
      description:
        type: string
      id:
        type: string
      receiver_bank_account_number:
        type: string
      receiver_bank_name:
        type: string
      receiver_id:
        type: string
      receiver_name:
        type: string
      sender_bank_account_number:
        type: string
      sender_bank_name:
        type: string
      sender_id:
        type: string
      sender_name:
        type: string
      source_transaction_id:
        type: string
      status:
        $ref: '#/definitions/transaction.Status'
      transaction_type:
        $ref: '#/definitions/transaction.TransactionType'
      update_time:
        type: string
    type: object
...
  transaction.Status:
    enum:
    - draft
    - draft
    - verified
    - success
    type: string
    x-enum-varnames:
    - DefaultStatus
    - StatusDraft
    - StatusVerified
    - StatusSuccess
  transaction.TransactionType:
    enum:
    - internal
    - external
    type: string
    x-enum-varnames:
    - TransactionTypeInternal
    - TransactionTypeExternal
info:
  contact: {}
paths:
...
  /api/customer/v1/customers/{id}:
    get:
      consumes:
      - application/json
      description: Get a customer
      operationId: customer-get
      parameters:
      - description: ID of customer
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.meResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Get a customer
      tags:
      - Customer
...
  /api/customer/v1/me/bank-accounts:
    get:
      consumes:
      - application/json
      description: Show bank accounts
      operationId: bankaccount-listing
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.EntitiesResponseTemplate-customer_bankAccountResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Show bank accounts
      tags:
      - Bank account
...
  /api/customer/v1/me/transactions:
    get:
      consumes:
      - application/json
      description: Show transactions
      operationId: transaction-listing
      parameters:
      - description: True if sort ascent by update_time otherwise ignored
        in: query
        name: update_time
        type: boolean
      - description: True if sort descent by update_time otherwise ignored
        in: query
        name: -update_time
        type: boolean
      - description: True if only debt transaction otherwise ignored
        in: query
        name: only_debt
        type: boolean
      - description: True if only receive transaction otherwise ignored
        in: query
        name: only_receive
        type: boolean
      - description: True if only send transaction otherwise ignored
        in: query
        name: only_send
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.EntitiesResponseTemplate-customer_transactionResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Show transactions
      tags:
      - Transaction
    post:
      consumes:
      - application/json
      description: Create a transaction
      operationId: transaction-create
      parameters:
      - description: Create a transaction
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/customer.transactionCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/customer.transactionCreateResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Create a transaction
      tags:
      - Transaction
  /api/customer/v1/me/transactions/{id}:
    get:
      consumes:
      - application/json
      description: Get a transaction
      operationId: transaction-get
      parameters:
      - description: ID of transaction
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.transactionResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Get a transaction
      tags:
      - Transaction
...
  /me/bank-accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a bank account
      operationId: bankaccount-delete
      parameters:
      - description: ID of bank account
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Delete a bank account
      tags:
      - Bank account
    get:
      consumes:
      - application/json
      description: Get a bank account
      operationId: bankaccount-get
      parameters:
      - description: ID of bank account
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.bankAccountResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Get a bank account
      tags:
      - Bank account
    put:
      consumes:
      - application/json
      description: Update a bank account
      operationId: bankaccount-update
      parameters:
      - description: Update a bank account
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/customer.bankAccountUpdateReq'
      - description: ID of bank account
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.bankAccountResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Update a bank account
      tags:
      - Bank account
...
swagger: "2.0"